Refactored model package
