+++ src/model/Floater.java
@@ -7,11 +7,10 @@ import util.*;
  * @author Model-team 
  * @version 1.0.0.0
  */
-public class Floater extends WorldObject{
-	protected Vector2D velocity;
+public class Floater extends Moveable implements Gravity{
+	protected final static double GRAVITY=1;
 	public Floater(double xSpeed, double ySpeed, double xPos, double yPos, double mass){
-		super(xPos, yPos, mass);
-		velocity = new Vector2D(xSpeed, ySpeed);
+		super(xSpeed, ySpeed, xPos, yPos, mass);
 	}
 	/**
 	 * Adds gravitational force to this Floater velocity
@@ -19,24 +18,37 @@ public class Floater extends WorldObject{
 	 * @param Affecting WorldObject
 	 * @return	void
 	 */
-	@Override
 	public void gravityPull(WorldCollection data) {
 		for(WorldObject obj : data.getCollection()){
-			if(!obj.equals(this))
-				velocity = velocity.add(this.calcuateGravity(obj));
+			if(!obj.equals(this) && obj instanceof Moveable)
+				((Moveable)obj).velocity =((Moveable)obj).velocity.add(this.calcuateGravity(obj));
 		}
 	}
 	/**
-	 * Move this Floater and notify observers
+	 * Calculates the scaled gravitational vector between two WorldObjects.
 	 * 
-	 * @param dt - delta time
-	 * @return
+	 * @param WorldObject
+	 * @return Vector2D the scaled gravity vector
 	 */
-	public void move(){
-		position = position.add(velocity);
-		System.out.println(position.toString());
-		update();
+	public Vector2D calcuateGravity(WorldObject arg) {
+		Vector2D temp = Vector2D.diff(this.getPosition(), arg.getPosition());
+		temp = temp.normalize();
+		temp = temp.scale(mass*GRAVITY*arg.getMass());
+		temp = temp.div(this.getPosition().distancesquared(arg.getPosition()));
+		return temp;
 	}
+	/**
+	 * Return this objects Gravity.
+	 * 
+	 * @return Double the gravity.
+	 */
+	public double getGravity() {
+		return GRAVITY;
+	}
+	/**
+	 * Return this objects velocity vector.
+	 * @return Vector2D the velocity vector.
+	 */
 	public Vector2D getVelocity(){
 		return velocity;
 	}
